Try using UNIX sockets instead
This fix should not be used as-is.  It's a proof-of-concept for
https://github.com/thoughtbot/capybara-webkit/issues/41 (Tests running
very slow on Linux).

I was seeing a delay of ~0.04 seconds on every command sent via TCP
waiting for webkit_server to come out of its poll() call.  So, perhaps
QtTcpServer has poor poll() performance.

Changed QtTcpServer and QtTcpSocket to QtLocalServer and QtLocalSocket.

Now tests that previously took my OpenSUSE machine ~26 minutes to run
take ~13 seconds.